tables: split partiton test
