+++ executor/BUILD.bazel
@@ -458,7 +458,6 @@ go_test(
         "@com_github_pingcap_kvproto//pkg/metapb",
         "@com_github_pingcap_log//:log",
         "@com_github_pingcap_sysutil//:sysutil",
-        "@com_github_pingcap_tipb//go-binlog",
         "@com_github_pingcap_tipb//go-tipb",
         "@com_github_prometheus_client_golang//prometheus",
         "@com_github_prometheus_client_model//go",
