+++ table/tables/BUILD.bazel
@@ -9,6 +9,7 @@ go_library(
         "partition.go",
         "state_remote.go",
         "tables.go",
+        "testutil.go",
     ],
     importpath = "github.com/pingcap/tidb/table/tables",
     visibility = ["//visibility:public"],
@@ -64,17 +65,15 @@ go_test(
     timeout = "long",
     srcs = [
         "cache_test.go",
-        "export_test.go",
         "index_test.go",
         "main_test.go",
         "mutation_checker_test.go",
-        "partition_test.go",
         "state_remote_test.go",
         "tables_test.go",
     ],
     embed = [":tables"],
     flaky = True,
-    shard_count = 50,
+    shard_count = 36,
     deps = [
         "//ddl",
         "//ddl/util/callback",
@@ -91,7 +90,6 @@ go_test(
         "//parser/mysql",
         "//session",
         "//sessionctx",
-        "//sessionctx/binloginfo",
         "//sessionctx/stmtctx",
         "//sessionctx/variable",
         "//sessiontxn",
@@ -105,20 +103,15 @@ go_test(
         "//util",
         "//util/codec",
         "//util/collate",
-        "//util/dbterror",
-        "//util/logutil",
         "//util/mock",
         "//util/rowcodec",
         "//util/stmtsummary",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
-        "@com_github_pingcap_tipb//go-binlog",
         "@com_github_prometheus_client_model//go",
         "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
         "@com_github_tikv_client_go_v2//oracle",
-        "@org_golang_google_grpc//:grpc",
         "@org_uber_go_goleak//:goleak",
-        "@org_uber_go_zap//:zap",
     ],
 )
